pyproject.toml
[project]
name = "tesci"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "requests",
    "pandas",
    "pyyaml",
    "PyGithub>=2.1.1",
    "attrs",
    "xlrd",
    "openpyxl",
    "rapidfuzz>=3.5.2",
    'importlib-metadata; python_version<"3.8"',
]

[project.scripts]
tesci = "tesci.scripts.tesci:cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages]
find = {} # Scan the project directory with the default parameters
